requirements.txt
yfinance
pandas
mysql-connector-python
streamlit
plotly
ta
prettytable
colorama